contact_role.go: Move file to package `common`
